Initial reorganization of BERT to split out dataset, model, and BERT-CoLA training example.
